feedback-backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
python-dotenv
flask-cors
